+++ backend/src/meal_shield/ranking/ranking_count.py
@@ -1,0 +1,10 @@
+import logging
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+# TODO: Make rule base count score func
+
+
+def score_allergens(recipes, allergens) -> dict[str, list[float]]:
+    pass
